requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
google-genai
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0
